Properly merge Sijax.request data request parameter
Reported in spantaleev/flask-sijax#6, stating that
a `data` object passed as an "additional request parameter" will
not get merged, but will be discarded in favor of Sijax.request's
own `data` object.

This makes it so that Sijax.request will start off from an empty
`data` object or from a user-provided one, and will just add in to that.